adopted image upload to version 1.0
