+++ djng/styling/bootstrap3/widgets.py
@@ -8,7 +8,9 @@ from django.utils.html import format_html
 from django.utils.safestring import mark_safe
 from django.utils.encoding import force_text
 from django.forms import widgets
-from djng.forms.widgets import (flatatt,
+from django.forms.utils import flatatt
+
+from djng.forms.widgets import (
     ChoiceFieldRenderer as DjngChoiceFieldRenderer, CheckboxChoiceInput as DjngCheckboxChoiceInput,
     CheckboxFieldRendererMixin, CheckboxSelectMultiple as DjngCheckboxSelectMultiple,
     RadioFieldRendererMixin, RadioSelect as DjngRadioSelect)
